+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Drop configured initialMaxRecords for favourited toots to half of the other configuration because it seems to be the bottleneck
 
 ## v0.60.0
 * (Demo App) Add a 'Create New Toot' button to the main page